refactor: 좋아요, 북마크 도메인 리팩토링 issue: SCRUM-199
