+++ src/modules/blogs/blog.service.ts
@@ -3,7 +3,7 @@ import { BlogRepository } from './repository/blog.repository';
 import { BlogEntity } from './entities/blog.entity';
 import { GetBlogResponse } from './dto/response/get.blog.response';
 import { GetBlogsQueryRequest } from './dto/request/get.blog.query.request';
-import { PaginationQueryDto } from '../../global/patination/pagination.query.dto';
+import { PaginationQueryDto } from '../../global/pagination/pagination.query.dto';
 
 @Injectable()
 export class BlogService {
