+++ src/modules/blogs/test/mock-data.ts
@@ -1,7 +1,7 @@
 import { BlogEntity } from '../entities/blog.entity';
 import { GetBlogResponse } from '../dto/response/get.blog.response';
 import { GetBlogsQueryRequest } from '../dto/request/get.blog.query.request';
-import { PaginationQueryDto } from '../../../global/patination/pagination.query.dto';
+import { PaginationQueryDto } from '../../../global/pagination/pagination.query.dto';
 import { BlogCategory } from '@prisma/client';
 
 export const blogEntity = (overrides?: Partial<BlogEntity>): BlogEntity => {
