+++ src/modules/bookmarks/dto/request/create.bookmark.request.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsBoolean, IsEnum, IsNumber } from 'class-validator';
 import { Type } from 'class-transformer';
-import { ContentCategory } from '../../../../global/common/category/content-category';
+import { ContentCategory } from '../../../../global/category/contentCategory';
 
 export class CreateBookmarkRequest {
     @IsNumber()
