+++ src/modules/likes/dto/request/create.like.request.ts
@@ -1,7 +1,7 @@
 import { IsBoolean, IsEnum, IsNumber } from 'class-validator';
 import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
-import { ContentCategory } from '../../../../global/common/category/content-category';
+import { ContentCategory } from '../../../../global/category/contentCategory';
 
 export class CreateLikeRequest {
     @IsNumber()
